Extract version template function into a version Jinja2 extension
