+++ homeassistant/helpers/template/__init__.py
@@ -21,7 +21,6 @@ from types import CodeType
 from typing import TYPE_CHECKING, Any, Concatenate, Literal, NoReturn, Self, overload
 import weakref
 
-from awesomeversion import AwesomeVersion
 import jinja2
 from jinja2 import pass_context, pass_eval_context
 from jinja2.runtime import AsyncLoopContext, LoopContext
@@ -1488,11 +1487,6 @@ def as_function(macro: jinja2.runtime.Macro) -> Callable[..., Any]:
     return wrapper
 
 
-def version(value):
-    """Filter and function to get version object of the value."""
-    return AwesomeVersion(value)
-
-
 def merge_response(value: ServiceResponse) -> list[Any]:
     """Merge action responses into single list.
 
@@ -1765,6 +1759,9 @@ class TemplateEnvironment(ImmutableSandboxedEnvironment):
         self.add_extension(
             "homeassistant.helpers.template.extensions.TypeCastExtension"
         )
+        self.add_extension(
+            "homeassistant.helpers.template.extensions.VersionExtension"
+        )
 
         self.globals["apply"] = apply
         self.globals["as_function"] = as_function
@@ -1772,7 +1769,6 @@ class TemplateEnvironment(ImmutableSandboxedEnvironment):
         self.globals["iif"] = iif
         self.globals["merge_response"] = merge_response
         self.globals["typeof"] = typeof
-        self.globals["version"] = version
         self.globals["zip"] = zip
 
         self.filters["add"] = add
@@ -1787,7 +1783,6 @@ class TemplateEnvironment(ImmutableSandboxedEnvironment):
         self.filters["random"] = random_every_time
         self.filters["round"] = forgiving_round
         self.filters["typeof"] = typeof
-        self.filters["version"] = version
 
         self.tests["apply"] = apply
         self.tests["contains"] = contains
