+++ tests/helpers/template/test_init.py
@@ -1065,24 +1065,6 @@ def test_timedelta(mock_is_safe, hass: HomeAssistant) -> None:
         assert result == "15 days"
 
 
-def test_version(hass: HomeAssistant) -> None:
-    """Test version filter and function."""
-    filter_result = render(hass, "{{ '2099.9.9' | version}}")
-    function_result = render(hass, "{{ version('2099.9.9')}}")
-    assert filter_result == function_result == "2099.9.9"
-
-    filter_result = render(hass, "{{ '2099.9.9' | version < '2099.9.10' }}")
-    function_result = render(hass, "{{ version('2099.9.9') < '2099.9.10' }}")
-    assert filter_result is function_result is True
-
-    filter_result = render(hass, "{{ '2099.9.9' | version == '2099.9.9' }}")
-    function_result = render(hass, "{{ version('2099.9.9') == '2099.9.9' }}")
-    assert filter_result is function_result is True
-
-    with pytest.raises(TemplateError):
-        render(hass, "{{ version(None) < '2099.9.10' }}")
-
-
 def test_distance_function_with_1_state(hass: HomeAssistant) -> None:
     """Test distance function with 1 state."""
     _set_up_units(hass)
